+++ src/Wallet/IFusionManager.h
@@ -31,7 +31,7 @@ public:
 
   virtual ~IFusionManager() {}
 
-  virtual size_t createFusionTransaction(uint64_t threshold, uint64_t mixin,
+  virtual size_t createFusionTransaction(uint64_t threshold, uint16_t mixin,
     const std::vector<std::string>& sourceAddresses = {}, const std::string& destinationAddress = "") = 0;
   virtual bool isFusionTransaction(size_t transactionId) const = 0;
   virtual EstimateResult estimate(uint64_t threshold, const std::vector<std::string>& sourceAddresses = {}) const = 0;
