+++ tests/TransfersTests/TestNodeRpcProxy.cpp
@@ -87,8 +87,8 @@ namespace {
     ASSERT_TRUE(mineBlocks(*nodeDaemons[NODE_0], wallet1.address(), 1));
     ASSERT_TRUE(mineBlocks(*nodeDaemons[NODE_0], wallet1.address(), m_currency.minedMoneyUnlockWindow()));
 
-    wallet1.waitForSynchronizationToHeight(static_cast<uint32_t>(m_currency.minedMoneyUnlockWindow()) + 1);
-    wallet2.waitForSynchronizationToHeight(static_cast<uint32_t>(m_currency.minedMoneyUnlockWindow()) + 1);
+    wallet1.waitForSynchronizationToHeight(m_currency.minedMoneyUnlockWindow() + 1);
+    wallet2.waitForSynchronizationToHeight(m_currency.minedMoneyUnlockWindow() + 1);
 
     PoolChangedObserver observer;
     node0->addObserver(&observer);
